CDAP-6385 - Secure Store passwords should be in cdap-security.xml
Added a "none" option for secure store provider. Setting this option will load a dummy class and performing any operations on that will display a message informing the user of how to configure secure store. This is now the default mode.
Updated the Guice module to return FileSecureStore or KMSSecureStore ony in valid cases.